% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/again.R
\name{again}
\alias{again}
\title{Ask again}
\usage{
again(
  conversation = last_conversation(),
  model = NULL,
  seed = NULL,
  temperature = NULL,
  top_p = NULL,
  cache = getOption("ask.cache"),
  api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{conversation}{A conversation, initiated by \code{ask()} or followed up by
\code{follow_up()}}

\item{model, seed, temperature, top_p}{inherited from the last item
of \code{conversation} by default}

\item{cache}{A path where to cache the outputs, or "ram" to store them
in RAM. useful to spare tokens and to have reproducible code.}

\item{api_key}{API key}
}
\description{
Ask the same thing again, by default with the same parameters, if \code{cache}
is used a new request will be sent for the given arguments, and cached
instead of the older value.
}
