% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_models.R
\name{all_models}
\alias{all_models}
\title{List OpenAI current models}
\usage{
all_models(api_key = Sys.getenv("OPENAI_API_KEY"))
}
\arguments{
\item{api_key}{API key}
}
\value{
A tibble
}
\description{
List OpenAI current models
}
