% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask-options.R
\name{ask-options}
\alias{ask-options}
\alias{ask.cache}
\title{Options for the 'ask' package}
\description{
\itemize{
\item \code{ask.cache}: For optimal reproducibility, a path to a directory where to
store the cache, or \code{ram} to store it into the RAM.
}
}
