% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask-package.R
\docType{package}
\name{ask-package}
\alias{ask-package}
\title{ask: ask R anything}
\description{
Interface to do useful things in R with LLMs.
}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
