% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{ask_}
\alias{ask_}
\title{Ask a series of questions, stored in an object}
\usage{
ask_(questions, .prompt = yellow(paste0(symbol$pointer, " ")))
}
\arguments{
\item{questions}{Questions stored with \code{\link{questions}}.}

\item{.prompt}{Prompt to prepend to all questions.}
}
\value{
A named list with the answers, see \code{\link{ask}}
  for the format.
}
\description{
Store questions with \code{\link{questions}}, and then ask them
with \code{ask_}.
}
\examples{
\dontrun{
qs <- questions(
  name = input("What is your name?"),
  color = choose("Red or blue?", c("Red", "Blue"))
)
ask_(qs)
}
}

