% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_boolean.R, R/ask_numeric.R
\name{ask_boolean}
\alias{ask_boolean}
\alias{ask_numeric}
\title{Ask a boolean question}
\usage{
ask_boolean(prompt = listen(), context = NULL, ...)

ask_numeric(prompt = listen(), unit = NULL, context = NULL, ...)
}
\arguments{
\item{prompt}{Your request, a string or a character vector that will be
concatenated to a string with line breaks as separators.}

\item{context}{An object of class "ask_context" usually built from a call
to \code{context()} or a \verb{context_*()} function. It is used to define a "system"
message that define the behavior, tone or focus of the assistant.}

\item{...}{forwarded to \code{ask()}}

\item{unit}{Optional, if not the SI is used, but better provide it.}
}
\description{
Ask a boolean question
}
