% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_in_place.R
\name{ask_in_place}
\alias{ask_in_place}
\title{Ask for script updates in place}
\usage{
ask_in_place(content = listen(), context = NULL, ...)
}
\arguments{
\item{content}{Your request: a string, or a vector that will be concatenated to a string with
line breaks as separators.}

\item{context}{An object of class "ask_context" usually built from a call
to \code{context()} or a \verb{context_*()} function. It is used to define a "system"
message that define the behavior, tone or focus of the assistant.}

\item{...}{Forwarded to \code{ask()}}
}
\description{
This will change files in place a terminal command in a new terminal, and print extra
information in the R console.
}
\examples{
\dontrun{
ask_in_place("update the existing readme with useful missing info", context = context_repo())
ask_terminal("show the latest 5 changes in compact form")
}
}
