% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_script.R
\name{ask_script}
\alias{ask_script}
\title{Ask in the context of a script}
\usage{
ask_script(content = listen(), file = NULL, context = NULL, ...)
}
\arguments{
\item{content}{Your request: a string, or a vector that will be concatenated to a string with
line breaks as separators.}

\item{file}{path to a script, if left NULL the active script is considered}

\item{context}{An object of class "ask_context" usually built from a call
to \code{context()} or a \verb{context_*()} function. It is used to define a "system"
message that define the behavior, tone or focus of the assistant.}

\item{...}{forwarded to \code{ask()}}
}
\description{
Ask in the context of a script
}
