% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_terminal.R
\name{ask_terminal}
\alias{ask_terminal}
\title{Ask for a terminal command}
\usage{
ask_terminal(content = listen(), context = NULL, ...)
}
\arguments{
\item{content}{Your request: a string, or a vector that will be concatenated to a string with
line breaks as separators.}

\item{context}{An object of class "ask_context" usually built from a call
to \code{context()} or a \verb{context_*()} function. It is used to define a "system"
message that define the behavior, tone or focus of the assistant.}

\item{...}{Forwarded to \code{ask()}}
}
\description{
This will suggest a terminal command in a new terminal, and print extra
information in the R console.
}
\examples{
\dontrun{
ask_terminal("who contributed to this project?")
ask_terminal("show the latest 5 changes in compact form")
}
}
