% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_plot.R
\name{capture_plot}
\alias{capture_plot}
\title{Capture plot}
\usage{
capture_plot(width = 800, height = 600)
}
\arguments{
\item{width}{image width}

\item{height}{image height}
}
\value{
The path to the created file
}
\description{
Capture current plot (can be a base plot, a ggplot2 plot or anything else)
and saves it into a temp png file whose path is returned. meant to be used
with the \code{image} argument of \verb{ask*()} functions.
}
