% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_clickup.R
\name{context_clickup}
\alias{context_clickup}
\title{Contextualize ClickUp data}
\usage{
context_clickup(
  team_id = Sys.getenv("CLICKUP_TEAM_ID"),
  api_key = Sys.getenv("CLICKUP_API_TOKEN"),
  num_tasks = 5
)
}
\arguments{
\item{team_id}{The ID of the ClickUp team for which to fetch data.}

\item{api_key}{API key for ClickUp.}

\item{num_tasks}{The number of recent tasks to fetch for each list (default is 5).}
}
\value{
An object of class "ask_context"
}
\description{
Fetch data from ClickUp and build a context using the \code{context()} function.
}
