% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_clipboard.R
\name{context_clipboard}
\alias{context_clipboard}
\title{Contextualize clipboard content}
\usage{
context_clipboard()
}
\value{
An object of class "ask_context"
}
\description{
Contextualize clipboard content
}
