% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexts.R
\name{context_commits}
\alias{context_commits}
\title{Contextualize git commits}
\usage{
context_commits(n = 5)
}
\arguments{
\item{n}{The number of commits to return context for (default is 5)}
}
\value{
An object of class "ask_context"
}
\description{
Contextualize git commits
}
