% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_github.R
\name{context_github}
\alias{context_github}
\title{Contextualize GitHub issues}
\usage{
context_github(
  repo = NULL,
  state = c("open", "closed", "all"),
  filter = c("assigned", "created", "mentioned", "subscribed", "all"),
  labels = NULL,
  sort = c("created", "updated", "comments"),
  direction = c("asc", "desc"),
  since = NULL,
  n = 30
)
}
\arguments{
\item{repo}{A string in the form 'my.org/my.repo' specifying the repository.
If not provided, will be fetched from the URL field of the DESCRIPTION file.}

\item{state}{State of the issues to return ('open', 'closed', or 'all').}

\item{filter}{A string, see details.}

\item{labels}{A vector of labels to filter the issues by.}

\item{sort}{What to sort the results by}

\item{direction}{The direction to sort the results by.}

\item{since}{A date or time used to only show results that were last updated after the given time.}

\item{n}{Number of results (max 100).}
}
\value{
An object of class "ask_context"
}
\description{
Contextualize GitHub issues
}
\section{values for filter}{

\itemize{
\item \strong{assigned} (default): Issues assigned to the authenticated user.
\item \strong{created}: Issues created by the authenticated user.
\item \strong{mentioned}: Issues mentioning the authenticated user.
\item \strong{subscribed}: Issues the user is subscribed to.
\item \strong{all}: All issues owned by the authenticated user.
}
}

