% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexts.R
\name{context_numeric}
\alias{context_numeric}
\title{Request a numeric output}
\usage{
context_numeric(unit = NULL)
}
\arguments{
\item{unit}{unit}
}
\value{
An object of class "ask_context"
}
\description{
Request a numeric output
}
