% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_object.R
\name{context_objects}
\alias{context_objects}
\title{Contextualize R objects}
\usage{
context_objects(objects)
}
\arguments{
\item{objects}{A list of R objects to be contextualized.}
}
\value{
An object of class "ask_context"
}
\description{
Contextualize R objects
}
