% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_package.R
\name{context_package}
\alias{context_package}
\title{Contextualize package}
\usage{
context_package(
  pkg,
  DESCRIPTION = TRUE,
  NAMESPACE = TRUE,
  README = TRUE,
  NEWS = TRUE,
  LICENSE = TRUE,
  help_files = TRUE,
  code = FALSE,
  vignettes = FALSE
)
}
\arguments{
\item{pkg}{Name of the package.}

\item{DESCRIPTION}{Boolean, whether to include the DESCRIPTION file.}

\item{NAMESPACE}{Boolean, whether to include the NAMESPACE file.}

\item{README}{Boolean, whether to include the README file.}

\item{NEWS}{Boolean, whether to include the NEWS.md file.}

\item{LICENSE}{Boolean, whether to include the LICENSE file.}

\item{help_files}{Boolean, whether to include help files.}

\item{code}{Boolean, whether to include code files.}

\item{vignettes}{Boolean, whether to include vignettes.}
}
\value{
An object of class "ask_context"
}
\description{
Contextualize package
}
