% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_pdf.R
\name{context_pdf}
\alias{context_pdf}
\title{Contextualize pdf text content}
\usage{
context_pdf(
  file,
  pages = NULL,
  opw = "",
  upw = "",
  dpi = 600,
  language = "eng",
  options = NULL,
  type = "text"
)
}
\arguments{
\item{file}{file path or raw vector with pdf data}

\item{pages}{which pages of the pdf file to extract}

\item{opw}{string with owner password to open pdf}

\item{upw}{string with user password to open pdf}

\item{dpi}{resolution to render image that is passed to \link[pdftools]{pdf_convert}.}

\item{language}{passed to \link[tesseract:tesseract]{tesseract} to specify the
languge of the engine.}

\item{options}{passed to \link[tesseract:tesseract]{tesseract} to specify OCR parameters}

\item{type}{one or more from "text", "ocr_text", "ocr_data". Decides wether to
use \code{pdftools::pdf_text()}, \code{pdftools::pdf_ocr_text()}, or/and \code{pdftools::pdf_ocr_data()}.}
}
\value{
An object of class "ask_context"
}
\description{
Contextualize pdf text content
}
\examples{
\dontrun{
ask("what is this about?", context_pdf("http://arxiv.org/pdf/1403.2805.pdf"))
}
}
