% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexts.R
\name{context_repo}
\alias{context_repo}
\title{Contextualize all R files in the repository}
\usage{
context_repo()
}
\value{
An object of class "ask_context"
}
\description{
Contextualize all R files in the repository
}
