% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexts.R
\name{context_script}
\alias{context_script}
\title{Contextualize a script}
\usage{
context_script(file = NULL)
}
\arguments{
\item{file}{Path to the file, if \code{NULL} we consider the active script}
}
\value{
An object of class "ask_context"
}
\description{
Contextualize a script
}
