% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexts.R
\name{context_session_info}
\alias{context_session_info}
\title{Contextualize the R session information}
\usage{
context_session_info()
}
\value{
An object of class "ask_context"
}
\description{
Contextualize the R session information
}
