% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexts.R
\name{context_url}
\alias{context_url}
\title{Contextualize an url}
\usage{
context_url(url, only_text = TRUE)
}
\arguments{
\item{url}{An url}

\item{only_text}{Whether to cleanup the html and extract the text only}
}
\value{
An object of class "ask_context"
}
\description{
Contextualize an url
}
