% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/again.R, R/follow_up.R
\name{again}
\alias{again}
\alias{follow_up}
\title{follow up}
\usage{
again(
  conversation = last_conversation(),
  model = "gpt-4o-2024-08-06",
  seed = NULL,
  temperature = NULL,
  top_p = NULL,
  local_seed = NULL,
  api_key = Sys.getenv("OPENAI_API_KEY")
)

follow_up(
  content = listen(),
  conversation = last_conversation(),
  model = "gpt-4o-2024-08-06",
  seed = NULL,
  temperature = NULL,
  top_p = NULL,
  local_seed = NULL,
  api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{conversation}{A conversation, initiated by \code{ask()} or followed up by
\code{follow_up()}}

\item{model, seed, temperature, top_p}{inherited by default from the last item
of \code{conversation} by default}

\item{local_seed}{If used, the result of a query is memoised so it will be
constant in the session, and quick to fetch. This ensures true reproducibility. Can be
used without using \code{seed}}

\item{api_key}{API key}

\item{content}{Your request: a string, or a vector that will be concatenated to a string with
line breaks as separators.}
}
\value{
a conversation object
}
\description{
follow up
}
