% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{json_object}
\alias{json_object}
\alias{json_properties}
\alias{json_array}
\alias{json_string}
\alias{json_number}
\alias{json_boolean}
\alias{json_null}
\title{JSON schema helpers}
\usage{
json_object(
  properties = json_properties(placeholder = json_null()),
  description = NULL,
  required = names(properties),
  additionalProperties = FALSE
)

json_properties(...)

json_array(items, description = NULL)

json_string(description = NULL)

json_number(description = NULL)

json_boolean(description = NULL)

json_null(description = NULL)
}
\arguments{
\item{properties}{An object built with \code{json_properties()}}

\item{description}{A string. Mostly useful for atomic items but can be used in \code{json_array()}
or \code{json_object()} as well}

\item{required}{An optional character vector of names of required properties}

\item{additionalProperties}{Either a boolean to allow or disallow additional properties,
or a an object built with \code{json_properties()} to specify a format that additional
properties must conform to. In its absence it's considered \code{TRUE}.}

\item{..., items}{Named objects, either basic objects built by \code{json_number()}, \code{json_string()},
or \code{json_boolean()}, or recursive objects built by \code{json_array()} or \code{json_object()}}
}
\description{
These are helpers meant to be used recursively in the \code{parameters} argument
of \code{new_tool()}. These build JSON schema element, or more precisely R nested
lists that can be converted to proper JSON schemas by \code{jsonlite::toJSON()}.
We support only very basic features of JSON schemas, if these are not sufficient
for your requirements when buidimg tools, please open a ticket.
}
