% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listen.R
\name{listen}
\alias{listen}
\title{listen to voice input}
\usage{
listen(verbose = TRUE)
}
\arguments{
\item{verbose}{whether to print the recorded text as a message}
}
\value{
A string
}
\description{
listen to voice input
}
