% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{new_tool}
\alias{new_tool}
\title{Use tools to get structured output}
\usage{
new_tool(name, description, parameters = json_object())
}
\arguments{
\item{name}{A string}

\item{description}{A string}

\item{parameters}{An object defined with \code{json_object()}}
}
\description{
Use tools to get structured output
}
\examples{
ask:::ask_impl("click at the center of the screen, my resolution is 1024 x 800", tools = list(tool_click(), tool_screenshot()))
tool_screenshot <- new_tool(
  "screenshot",
  "Instruct the user to take a screenshot",
 )
new_tool(
  "mac_terminal",
  "Provide the macOS terminal command that will fulfill the request. It can be used for instance to run an app or interrogate the system",
  command = "The terminal command",
  text = "A comment on why we use the given command"
)
context_computer <- "you'll iterate with the user using the tools at your disposal to request for screenshots, clicks etc, to browse the internet you can open a browser through the terminal, take screenshots and guide the user"
ask:::ask_impl("what app is my screen showing now ?", context_computer, tools = tools_openai)
response_data(last_conversation()$response[[1]]) |> constructive::construct()
ask:::ask_impl("Find a picture of a poodle online for me", context_computer, tools = tools_openai)
new_claude_tool(
  "mac_terminal",
  "Provide the macOS terminal command that will fulfill the request. It can be used for instance to run an app or interrogate the system",
  command = "The terminal command"
)
}
