% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ask.R
\name{questions}
\alias{questions}
\title{Store a series of questions, to ask them later}
\usage{
questions(...)
}
\arguments{
\item{...}{Questions to store. See \code{\link{ask}}.}
}
\value{
An unevaluated series of questions for future use.
}
\description{
Later you can call \code{\link{ask_}} (note the trailing underscore!)
to ask them.
}
\examples{
\dontrun{
qs <- questions(
  name = input("What is your name?"),
  color = choose("Red or blue?", c("Red", "Blue"))
)
ask_(qs)
}
}

