% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_models.R
\name{all_models}
\alias{all_models}
\title{List all current models}
\usage{
all_models(
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  only_supported = TRUE
)
}
\arguments{
\item{openai_api_key}{an api key for open ai (for other models we use
webscraping rather than rest APIs to get the list of models)}
}
\value{
A tibble
}
\description{
These are all current models from openai, anthropic and ollama. We don't
guarantee that they all work with ask though as we've tested a limited amount.
ChatGPT, Claude, and llama models seem to work ok, please open a ticket
to request further support.
}
