% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask-package.R
\docType{package}
\name{ask-package}
\alias{ask-package}
\title{ask: ask R anything}
\description{
Interface to do useful things in R with LLMs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moodymudskipper/ask}
  \item Report bugs at \url{https://github.com/moodymudskipper/ask/issues}
}

}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

Other contributors:
\itemize{
  \item Jordi Rosell \email{jroselln@gmail.com} [contributor]
}

}
\keyword{internal}
