% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ask.R
\name{ask}
\alias{ask}
\title{Ask a question to the user, through the command line}
\usage{
ask(...)
}
\arguments{
\item{...}{Questions to ask, see details below.}
}
\value{
A named list with the answers.
}
\description{
Ask a question to the user, through the command line
}
\examples{
\dontrun{
ask(
  name =  input("What is your name?"),
  cool = confirm("Are you cool?"),
  drink = choose("Select your poison!", c("Beer", "Wine"))
)
}
}

