% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{ask}
\alias{ask}
\title{Ask chat gpt anything}
\usage{
ask(
  content = listen(),
  context = NULL,
  model = "gpt-4o-2024-08-06",
  seed = NULL,
  temperature = 1,
  top_p = 1,
  cache = getOption("ask.cache"),
  api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{content}{Your request: a string, or a vector that will be concatenated to a string with
line breaks as separators.}

\item{context}{An object of class "ask_context" usually built from a call
to \code{context()} or a \verb{context_*()} function. It is used to define a "system"
message that define the behavior, tone or focus of the assistant.}

\item{model}{The model to choose, see https://platform.openai.com/docs/models
or call \code{all_models()}}

\item{seed}{The seed used by the model, makes things more reproducible, but
not completely, due to the nature of LLMs. See \code{cache} to work around
that.}

\item{temperature}{Choose higher \code{temperature} for more diverse and
unexpected results, and lower \code{temperature} for more controlled and
consistent text.}

\item{top_p}{Choose high \code{top_p} for creative applications like storytelling,
poetry, or brainstorming. Choose low \code{top_p} for applications requiring
precision and coherence, such as technical writing, factual content, or summarization.}

\item{cache}{A path where to cache the outputs, or "ram" to store them
in RAM. useful to spare tokens and to have reproducible code.}

\item{api_key}{API key}
}
\value{
a converstaion object
}
\description{
Ask chat gpt anything
}
