% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_document.R
\name{ask_script}
\alias{ask_script}
\title{Ask in the context of a script}
\usage{
ask_script(content = listen(), path = NULL, ...)
}
\arguments{
\item{content}{Your request: a string, or a vector that will be concatenated to a string with
line breaks as separators.}

\item{path}{path to a script, if left NULL the active script is considered}

\item{...}{forwarded to \code{ask()}}
}
\description{
Ask in the context of a script
}
