% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow_up.R
\name{follow_up}
\alias{follow_up}
\title{Follow up a request}
\usage{
follow_up(
  content = listen(),
  context = NULL,
  conversation = last_conversation(),
  model = "gpt-4o-2024-08-06",
  seed = NULL,
  temperature = NULL,
  top_p = NULL,
  cache = getOption("ask.cache"),
  api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{content}{Your request: a string, or a vector that will be concatenated to a string with
line breaks as separators.}

\item{context}{An object of class "ask_context" usually built from a call
to \code{context()} or a \verb{context_*()} function. It is used to define a "system"
message that define the behavior, tone or focus of the assistant.}

\item{conversation}{A conversation, initiated by \code{ask()} or followed up by
\code{follow_up()}}

\item{model, seed, temperature, top_p}{inherited from the last item
of \code{conversation} by default}

\item{cache}{A path where to cache the outputs, or "ram" to store them
in RAM. useful to spare tokens and to have reproducible code.}

\item{api_key}{API key}
}
\value{
a conversation object
}
\description{
Continue a conversation, by default with the same parameters.
}
